Дорабатывает и исправляет проект для отправки на проверку
