+++ js/gallery.js
@@ -1,6 +1,5 @@
 const photosContainerElement = document.querySelector('.pictures');
 const photoTemplate = document.querySelector('#picture').content;
-const imageFiltersElement = document.querySelector('.img-filters');
 
 const getPhotoElement = ({url, id, likes, comments}) => {
   const element = photoTemplate.cloneNode(true);
@@ -21,7 +20,6 @@ const addPhotos = (photos) => {
   });
 
   photosContainerElement.append(photosFragment);
-  imageFiltersElement.classList.remove('img-filters--inactive');
 };
 
-export {addPhotos, imageFiltersElement};
+export {addPhotos};
