+++ js/network.js
@@ -1,5 +1,3 @@
-import {showAlert, showMessageModal} from './message.js';
-
 const DATA_LOAD_ERROR = 'Для получения данных оплатите обращение к серверу';
 const KEKSTAGRAM_SERVER = 'https://26.javascript.pages.academy/kekstagram';
 
@@ -10,7 +8,7 @@ const getPhotos = (onSuccess, onFail) => {
       onSuccess(photos);
     })
     .catch(() => {
-      onFail(showAlert(DATA_LOAD_ERROR));
+      onFail(DATA_LOAD_ERROR);
     });
 };
 
@@ -25,13 +23,12 @@ const sendData = (onSuccess, onFail, body) => {
     .then((response) => {
       if (response.ok) {
         onSuccess();
-        showMessageModal('success');
       } else {
-        onFail(showMessageModal('error'));
+        onFail();
       }
     })
     .catch(() => {
-      onFail(showMessageModal('error'));
+      onFail();
     });
 };
